requirements.txt
requests==2.3.0
boto==2.28.0
importlib==1.0.3
python-daemon==1.6
beautifulsoup4==4.3.2
jira==0.25
